global refactoring
