+++ general/additional_funcs.h
@@ -1,0 +1,12 @@
+//
+// Created by Aletha Yellin on 20/11/2019.
+//
+
+#ifndef PRINTF_ADDITIONAL_FUNCS_H
+#define PRINTF_ADDITIONAL_FUNCS_H
+
+void	ft_putempty(const char *fmt, va_list args);
+
+void	ft_putpercent(const char *fmt, va_list args);
+
+#endif //PRINTF_ADDITIONAL_FUNCS_H
