+++ general/basic_funcs.h
@@ -1,0 +1,12 @@
+//
+// Created by Aletha Yellin on 20/11/2019.
+//
+
+#ifndef PRINTF_BASIC_FUNCS_H
+#define PRINTF_BASIC_FUNCS_H
+
+void	ft_putstr_args(const char *fmt, va_list args);
+
+void	ft_putnum_args(const char *fmt, va_list args);
+
+#endif //PRINTF_BASIC_FUNCS_H
