+++ general/format_parser.h
@@ -1,0 +1,10 @@
+//
+// Created by Aletha Yellin on 20/11/2019.
+//
+
+#ifndef PRINTF_FORMAT_PARSER_H
+#define PRINTF_FORMAT_PARSER_H
+
+int		get_format_length(const char *format);
+
+#endif //PRINTF_FORMAT_PARSER_H
