+++ general/ft_printf.h
@@ -1,0 +1,21 @@
+//
+// Created by Aletha Yellin on 06/11/2019.
+//
+
+#ifndef PRINTF_FT_PRINTF_H
+#define PRINTF_FT_PRINTF_H
+
+# include <stdarg.h>
+
+# include "libft.h"
+# include "printf_constatns.h"
+# include "basic_funcs.h"
+# include "additional_funcs.h"
+# include "format_parser.h"
+# include "util.h"
+# include "additional_funcs.h"
+# include "func_table.h"
+
+int		ft_printf(const char *format, ...);
+
+#endif //PRINTF_FT_PRINTF_H
