--- general/ft_stdio.h
@@ -1,12 +1,0 @@
-//
-// Created by Aletha Yellin on 06/11/2019.
-//
-
-#ifndef PRINTF_FT_STDIO_H
-#define PRINTF_FT_STDIO_H
-
-# include "libft.h"
-
-int		ft_printf(const char *format, ...);
-
-#endif //PRINTF_FT_STDIO_H
