+++ general/libft/ft_strcspn.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strcspn.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ayellin <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/11/20 17:46:09 by ayellin           #+#    #+#             */
+/*   Updated: 2019/11/20 17:46:11 by ayellin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+ssize_t	ft_strcspn(const char *s, const char *charset)
+{
+	ssize_t i;
+	ssize_t j;
+
+	i = 0;
+	while (s[i])
+	{
+		j = 0;
+		while (charset[j])
+		{
+			if (s[i] == charset[j])
+				return (i);
+			j++;
+		}
+		i++;
+	}
+	return (-1);
+}
