+++ general/libft/includes/libft.h
@@ -129,7 +129,7 @@ ssize_t			ft_strchri(const char *s, char c);
 void			ft_str_shift_left(char *str, size_t from);
 char			*ft_strjoinf(char *s1, char *s2, int flag);
 char			*ft_strsubf(char **ps, size_t start, size_t len);
-
+ssize_t			ft_strcspn(const char *s, const char *charset);
 /*
  **	Memory:
 */
