+++ general/printf_constatns.h
@@ -1,0 +1,16 @@
+//
+// Created by Aletha Yellin on 20/11/2019.
+//
+
+#ifndef PRINTF_PRINTF_CONSTATNS_H
+#define PRINTF_PRINTF_CONSTATNS_H
+
+#define FT_PRINTF_STR 1
+#define FT_PRINTF_D 2
+#define FT_PRINTF_PERCENT 3
+#define FT_PRINTF_EMPTY 0
+
+static const char *g_specificators = "cpsdiouxX";
+
+
+#endif //PRINTF_PRINTF_CONSTATNS_H
