+++ general/util.h
@@ -1,0 +1,17 @@
+//
+// Created by Aletha Yellin on 20/11/2019.
+//
+
+#ifndef PRINTF_UTIL_H
+#define PRINTF_UTIL_H
+
+typedef struct s_pair
+{
+	size_t x;
+	size_t y;
+}				t_pair;
+
+t_pair		*ft_strfind(const char *str, const char *charset);
+
+
+#endif //PRINTF_UTIL_H
